+++ packages/compiler/src/core/program.ts
@@ -1,3 +1,4 @@
+import pc from "picocolors";
 import { EmitterOptions } from "../config/types.js";
 import { setCurrentProgram } from "../experimental/typekit/index.js";
 import { validateEncodedNamesConflicts } from "../lib/encoded-names.js";
@@ -11,7 +12,8 @@ import {
   resolveModule,
 } from "../module-resolver/module-resolver.js";
 import { PackageJson } from "../types/package-json.js";
-import { deepEquals, findProjectRoot, isDefined, mapEquals, mutate } from "../utils/misc.js";
+import { findProjectRoot } from "../utils/io.js";
+import { deepEquals, isDefined, mapEquals, mutate } from "../utils/misc.js";
 import { createBinder } from "./binder.js";
 import { Checker, createChecker } from "./checker.js";
 import { createSuppressCodeFix } from "./compiler-code-fixes/suppress.codefix.js";
@@ -32,13 +34,7 @@ import { createDiagnostic } from "./messages.js";
 import { createResolver } from "./name-resolver.js";
 import { CompilerOptions } from "./options.js";
 import { parse, parseStandaloneTypeReference } from "./parser.js";
-import {
-  getDirectoryPath,
-  getRelativePathFromDirectory,
-  joinPaths,
-  resolvePath,
-} from "./path-utils.js";
-import { createProjector } from "./projector.js";
+import { getDirectoryPath, joinPaths, resolvePath } from "./path-utils.js";
 import {
   SourceLoader,
   SourceResolution,
@@ -46,7 +42,7 @@ import {
   loadJsFile,
   moduleResolutionErrorToDiagnostic,
 } from "./source-loader.js";
-import { StateMap, StateSet, createStateAccessors } from "./state-accessors.js";
+import { createStateAccessors } from "./state-accessors.js";
 import {
   CompilerHost,
   Diagnostic,
@@ -60,12 +56,11 @@ import {
   LibraryMetadata,
   LiteralType,
   LocationContext,
+  LogSink,
   ModuleLibraryMetadata,
   Namespace,
   NoTarget,
   Node,
-  ProjectionApplication,
-  Projector,
   SourceFile,
   Sym,
   SymbolFlags,
@@ -78,50 +73,50 @@ import {
   TypeSpecScriptNode,
 } from "./types.js";
 
-/** @deprecated */
-export interface ProjectedProgram extends Program {
-  projector: Projector;
-}
-
-/** @deprecated use Mutators instead */
-export function projectProgram(
-  program: Program,
-  projections: ProjectionApplication[],
-  startNode?: Type,
-  // eslint-disable-next-line @typescript-eslint/no-deprecated
-): ProjectedProgram {
-  return createProjector(program, projections, startNode);
-}
-
 export interface Program {
   compilerOptions: CompilerOptions;
+  /** @internal */
   mainFile?: TypeSpecScriptNode;
   /** All source files in the program, keyed by their file path. */
   sourceFiles: Map<string, TypeSpecScriptNode>;
   jsSourceFiles: Map<string, JsSourceFileNode>;
+
+  /** @internal */
   literalTypes: Map<string | number | boolean, LiteralType>;
   host: CompilerHost;
   tracer: Tracer;
   trace(area: string, message: string): void;
+  /**
+   * **DANGER** Using the checker is reserved for advanced usage and should be used with caution.
+   * API are not subject to the same stability guarantees see See https://typespec.io/docs/handbook/breaking-change-policy/
+   */
   checker: Checker;
   emitters: EmitterRef[];
   readonly diagnostics: readonly Diagnostic[];
+  /** @internal */
   loadTypeSpecScript(typespecScript: SourceFile): Promise<TypeSpecScriptNode>;
+  /** @internal */
   onValidate(
     cb: (program: Program) => void | Promise<void>,
     LibraryMetadata: LibraryMetadata,
   ): void;
+  /** @internal */
   getOption(key: string): string | undefined;
   stateSet(key: symbol): Set<Type>;
-  stateSets: Map<symbol, StateSet>;
+  /** @internal */
+  stateSets: Map<symbol, Set<Type>>;
   stateMap(key: symbol): Map<Type, any>;
-  stateMaps: Map<symbol, StateMap>;
+  /** @internal */
+  stateMaps: Map<symbol, Map<Type, unknown>>;
   hasError(): boolean;
   reportDiagnostic(diagnostic: Diagnostic): void;
   reportDiagnostics(diagnostics: readonly Diagnostic[]): void;
+
+  /** @internal */
   reportDuplicateSymbols(symbols: SymbolTable | undefined): void;
 
   getGlobalNamespaceType(): Namespace;
+
   resolveTypeReference(reference: string): [Type | undefined, readonly Diagnostic[]];
 
   /** Return location context of the given source file. */
@@ -139,6 +134,7 @@ interface EmitterRef {
   metadata: LibraryMetadata;
   emitterOutputDir: string;
   options: Record<string, unknown>;
+  readonly library: LibraryInstance;
 }
 
 interface Validator {
@@ -158,8 +154,18 @@ export async function compile(
   oldProgram?: Program, // NOTE: deliberately separate from options to avoid memory leak by chaining all old programs together.
 ) {
   const logger = createLogger({ sink: host.logSink });
-  const { program, shouldAbort } = await logger.trackAction("Compiling...", "Compiling", () =>
-    createProgram(host, mainFile, options, oldProgram),
+  const { program, shouldAbort } = await logger.trackAction(
+    "Compiling...",
+    "Compiling",
+    async (task) => {
+      const result = await createProgram(host, mainFile, options, oldProgram);
+      if (result.program.hasError()) {
+        task.fail();
+      } else if (result.program.diagnostics.length > 0) {
+        task.warn();
+      }
+      return result;
+    },
   );
 
   if (shouldAbort) {
@@ -168,10 +174,14 @@ export async function compile(
 
   // Emitter stage
   for (const emitter of program.emitters) {
-    await emit(emitter, program, options);
+    // If in dry mode run and an emitter doesn't support it we have to skip it.
+    if (program.compilerOptions.dryRun && !emitter.library.definition?.capabilities?.dryRun) {
+      continue;
+    }
+    await emit(emitter, program);
 
     if (options.listFiles) {
-      logEmittedFilesPath(program.projectRoot);
+      logEmittedFilesPath(host.logSink);
     }
   }
   return program;
@@ -184,8 +194,8 @@ async function createProgram(
   oldProgram?: Program,
 ): Promise<{ program: Program; shouldAbort: boolean }> {
   const validateCbs: Validator[] = [];
-  const stateMaps = new Map<symbol, StateMap>();
-  const stateSets = new Map<symbol, StateSet>();
+  const stateMaps = new Map<symbol, Map<Type, unknown>>();
+  const stateSets = new Map<symbol, Set<Type>>();
   const diagnostics: Diagnostic[] = [];
   const duplicateSymbols = new Set<Sym>();
   const emitters: EmitterRef[] = [];
@@ -243,10 +253,10 @@ async function createProgram(
 
   await loadSources(resolvedMain);
 
-  const emit = options.emit;
+  const emit = options.noEmit ? [] : (options.emit ?? []);
   const emitterOptions = options.options;
 
-  await loadEmitters(basedir, emit ?? [], emitterOptions ?? {});
+  await loadEmitters(basedir, emit, emitterOptions ?? {});
 
   if (
     oldProgram &&
@@ -539,6 +549,7 @@ async function createProgram(
         metadata,
         emitterOutputDir,
         options: emitterOptions,
+        library,
       };
     } else {
       program.trace(
@@ -907,17 +918,28 @@ function resolveOptions(options: CompilerOptions): CompilerOptions {
   return { ...options };
 }
 
-async function emit(emitter: EmitterRef, program: Program, options: CompilerOptions = {}) {
+async function emit(emitter: EmitterRef, program: Program) {
   const emitterName = emitter.metadata.name ?? "";
-  const relativePathForEmittedFiles = options.listFiles
-    ? `./${getRelativePathFromDirectory(program.projectRoot, emitter.emitterOutputDir, false)}/`
-    : "";
+  const relativePathForEmittedFiles =
+    transformPathForSink(program.host.logSink, emitter.emitterOutputDir) + "/";
 
+  const errorCount = program.diagnostics.filter((x) => x.severity === "error").length;
+  const warnCount = program.diagnostics.filter((x) => x.severity === "warning").length;
   const logger = createLogger({ sink: program.host.logSink });
   await logger.trackAction(
     `Running ${emitterName}...`,
-    `${emitterName}\t${relativePathForEmittedFiles}`,
-    () => runEmitter(emitter, program),
+    `${emitterName}    ${pc.dim(relativePathForEmittedFiles)}`,
+    async (task) => {
+      await runEmitter(emitter, program);
+
+      const newErrorCount = program.diagnostics.filter((x) => x.severity === "error").length;
+      const newWarnCount = program.diagnostics.filter((x) => x.severity === "warning").length;
+      if (newErrorCount > errorCount) {
+        task.fail();
+      } else if (newWarnCount > warnCount) {
+        task.warn();
+      }
+    },
   );
 }
 
@@ -937,9 +959,12 @@ async function runEmitter(emitter: EmitterRef, program: Program) {
   }
 }
 
-function logEmittedFilesPath(projectRoot: string) {
+function logEmittedFilesPath(logSink: LogSink) {
   flushEmittedFilesPaths().forEach((filePath) => {
     // eslint-disable-next-line no-console
-    console.log(`\t./${getRelativePathFromDirectory(projectRoot, filePath, false)}`);
+    console.log(`    ${pc.dim(transformPathForSink(logSink, filePath))}`);
   });
+}
+function transformPathForSink(logSink: LogSink, path: string) {
+  return logSink.getPath ? logSink.getPath(path) : path;
 }