+++ packages/compiler/src/init/init.ts
@@ -91,14 +91,25 @@ export async function initTypeSpecProjectWorker(
   whiteline();
 
   if (projectJsonCreated) {
-    await action(
+    await host.logSink.trackAction!(
       "Installing dependencies",
-      async () =>
-        await installTypeSpecDependencies(host, {
+      "Dependencies installed",
+      async (task) => {
+        const diagnostics = await installTypeSpecDependencies(host, {
           directory,
           stdio: "pipe",
           savePackageManager: true,
-        }),
+        });
+
+        if (diagnostics.length > 0) {
+          if (diagnostics.some((d) => d.severity === "error")) {
+            task.fail();
+          } else {
+            task.warn();
+          }
+          logDiagnostics(diagnostics);
+        }
+      },
     );
   }
 
@@ -361,57 +372,4 @@ function logDiagnostics(diagnostics: readonly Diagnostic[]): void {
     // eslint-disable-next-line no-console
     console.log(diagnostic.message);
   });
-}
-
-async function action(message: string, fn: () => Promise<unknown>) {
-  const isTTY = process.stdout?.isTTY && !process.env.CI;
-
-  let interval;
-
-  if (isTTY) {
-    const spinner = createSpinner();
-    interval = setInterval(() => {
-      process.stdout.clearLine(0);
-      process.stdout.cursorTo(0);
-      process.stdout.write(`\r  ${pc.yellow(spinner())} ${message}`);
-    }, 300);
-  } else {
-    // eslint-disable-next-line no-console
-    console.log(`  - ${message}`);
-  }
-
-  try {
-    await fn();
-    if (isTTY) {
-      process.stdout.clearLine(0);
-      process.stdout.cursorTo(0);
-      process.stdout.write(`\r  ${pc.green("✓")} ${message}\n`);
-    } else {
-      // eslint-disable-next-line no-console
-      console.log(`  ${pc.green("✓")} ${message}`);
-    }
-  } catch (e: any) {
-    // eslint-disable-next-line no-console
-    console.log(`\r  ${pc.red("×")} ${message}\n${e.message}`);
-    process.exit(1);
-  } finally {
-    if (interval) {
-      clearInterval(interval);
-    }
-  }
-}
-
-import isUnicodeSupported from "is-unicode-supported";
-
-export const spinnerFrames = isUnicodeSupported()
-  ? ["⠋", "⠙", "⠹", "⠸", "⠼", "⠴", "⠦", "⠧", "⠇", "⠏"]
-  : ["-", "\\", "|", "/"];
-
-export function createSpinner() {
-  let index = 0;
-
-  return () => {
-    index = ++index % spinnerFrames.length;
-    return spinnerFrames[index];
-  };
 }