+++ packages/typespec-vscode/src/utils.ts
@@ -1,6 +1,6 @@
 import type { ModuleResolutionResult, PackageJson, ResolveModuleHost } from "@typespec/compiler";
 import { spawn, SpawnOptions } from "child_process";
-import { readFile, realpath, stat } from "fs/promises";
+import { mkdtemp, readdir, readFile, realpath, stat } from "fs/promises";
 import { dirname } from "path";
 import { CancellationToken } from "vscode";
 import { Executable } from "vscode-languageclient/node.js";
@@ -27,6 +27,15 @@ export async function isDirectory(path: string) {
     return stats.isDirectory();
   } catch {
     return false;
+  }
+}
+
+export async function createTempDir(tmpRoot: string, prefix: string): Promise<string | undefined> {
+  try {
+    return await mkdtemp(joinPaths(tmpRoot, prefix));
+  } catch (e) {
+    logger.error("Failed to create temp folder", [e]);
+    return undefined;
   }
 }
 
@@ -125,6 +134,14 @@ export async function tryReadFile(path: string): Promise<string | undefined> {
     return content;
   } catch (e) {
     logger.debug(`Failed to read file: ${path}`, [e]);
+    return undefined;
+  }
+}
+
+export async function tryReadDir(path: string): Promise<string[] | undefined> {
+  try {
+    return await readdir(path);
+  } catch (e) {
     return undefined;
   }
 }
@@ -404,9 +421,51 @@ export async function loadPackageJsonFile(
  * @returns the path to the installed node executable, or empty string if not found.
  */
 export async function checkInstalledNode(): Promise<string> {
+  return checkInstalledExecutable("node");
+}
+
+export async function checkInstalledTspCli(): Promise<string> {
+  return checkInstalledExecutable("tsp");
+}
+
+export async function checkInstalledNpm(): Promise<string> {
+  return checkInstalledExecutable("npm");
+}
+
+export async function checkInstalledExecutable(exe: string): Promise<string> {
   try {
-    return await which("node");
+    return await which(exe);
   } catch (e) {
     return "";
   }
+}
+
+export async function parseJsonFromFile(filePath: string): Promise<string | undefined> {
+  try {
+    const fileContent = await readFile(filePath, "utf-8");
+    const content = JSON.parse(fileContent);
+    return content;
+  } catch (e) {
+    logger.error(`Failed to load JSON file: ${filePath}`, [e]);
+    return;
+  }
+}
+
+/**
+ * Throttle the function to be called at most once in every blockInMs milliseconds. This utility
+ * is useful when your event handler will trigger the same event multiple times in a short period.
+ *
+ * @param fn Underlying function to be throttled
+ * @param blockInMs Block time in milliseconds
+ * @returns a throttled function
+ */
+export function throttle<T extends (...args: any[]) => any>(fn: T, blockInMs: number): T {
+  let time: number | undefined;
+  return function (this: any, ...args: Parameters<T>) {
+    const now = Date.now();
+    if (time === undefined || now - time >= blockInMs) {
+      time = now;
+      fn.apply(this, args);
+    }
+  } as T;
 }